+++ ibm_mq/tests/common.py
@@ -5,7 +5,12 @@
 import os
 
 from datadog_checks.dev import get_docker_hostname
-from datadog_checks.ibm_mq.metrics import COUNT, GAUGE
+
+# Ignore missing library to not require it for e2e
+try:
+    from datadog_checks.ibm_mq.metrics import COUNT, GAUGE
+except ImportError:
+    COUNT = GAUGE = ''
 
 HERE = os.path.dirname(os.path.abspath(__file__))
 COMPOSE_DIR = os.path.join(HERE, 'compose')
@@ -175,7 +180,10 @@ QUEUE_METRICS = [
 ]
 
 QUEUE_STATUS_METRICS = [
+    ('ibm_mq.queue.oldest_message_age', GAUGE),
     ('ibm_mq.queue.uncommitted_msgs', GAUGE),
+    ('ibm_mq.queue.last_get_time', GAUGE),
+    ('ibm_mq.queue.last_put_time', GAUGE),
 ]
 
 CHANNEL_METRICS = [